+++ ipytest/_impl.py
@@ -11,15 +11,26 @@ import sys
 import tempfile
 import threading
 
+from typing import Any, Dict
+
 import packaging.version
 import pytest
 
 from IPython import get_ipython
 
-from ._config import current_config, default_clean
+from ._config import current_config, default
 
 
-def run(*args, module=None, plugins=()):
+def run(
+    *args,
+    module=None,
+    plugins=(),
+    run_in_thread=default,
+    raise_on_error=default,
+    addopts=default,
+    defopts=default,
+    display_columns=default,
+):
     """Execute all tests in the passed module (defaults to `__main__`) with pytest.
 
     **Parameters:**
@@ -27,27 +38,42 @@ def run(*args, module=None, plugins=()):
     - `args`: additional commandline options passed to pytest
     - `module`: the module containing the tests. If not given, `__main__` will
       be used.
-    - `filename`: the filename of the file containing the tests. It has to be a
-      real file, e.g., a notebook name, since its existence will be checked by
-      pytest. If not given, the `__file__` attribute of the passed module will
-      be used.
     - `plugins`: additional plugins passed to pytest.
+
+    The following parameters override the config options set with
+    [`ipytest.config()`][ipytest.config] or
+    [`ipytest.autoconfig()`][ipytest.autoconfig].
+
+    - `run_in_thread`: if given, override the config option "run_in_thread".
+    - `raise_on_error`: if given, override the config option "raise_on_error".
+    - `addopts`: if given, override the config option "addopts".
+    - `defopts`: if given, override the config option "defopts".
+    - `display_columns`: if given, override the config option "display_columns".
 
     **Returns**: the exit code of `pytest.main`.
     """
     import ipytest
 
-    run = run_in_thread if current_config["run_in_thread"] else run_direct
+    run_in_thread = default.unwrap(run_in_thread, current_config["run_in_thread"])
+    raise_on_error = default.unwrap(raise_on_error, current_config["raise_on_error"])
+    addopts = default.unwrap(addopts, current_config["addopts"])
+    defopts = default.unwrap(defopts, current_config["defopts"])
+    display_columns = default.unwrap(display_columns, current_config["display_columns"])
+
+    run = run_func_in_thread if run_in_thread else run_func_direct
     exit_code = run(
         _run_impl,
         *args,
         module=module,
         plugins=plugins,
+        addopts=addopts,
+        defopts=defopts,
+        display_columns=display_columns,
     )
 
     ipytest.exit_code = exit_code
 
-    if current_config["raise_on_error"] is True and exit_code != 0:
+    if raise_on_error is True and exit_code != 0:
         raise Error(exit_code)
 
     return exit_code
@@ -63,31 +89,44 @@ class Error(RuntimeError):
         return f"ipytest failed with exit_code {self.args[0]}"
 
 
-def pytest_magic(line, cell):
-    """IPython magic to execute pytest.
+def pytest_magic(line, cell, module=None):
+    """IPython magic to first execute the cell, then execute [`ipytest.run()`][ipytest.run].
 
-    It first executes the cell, then executes `ipytest.run()`. Any arguments
-    passed on the magic line be passed on to pytest. It cleans any previously
-    found tests, i.e., only tests defined in the current cell are executed. To
-    disable this behavior, use `ipytest.config(clean=False)`. To register the
-    magics, run `ipytest.autoconfig()` or `ipytest.config(magics=True)` first.
+    **Note:** the magics are only available after running
+    [`ipytest.autoconfig()`][ipytest.autoconfig] or
+    [`ipytest.config(magics=True)`][ipytest.config].
 
-    Additional arguments can be passed to pytest. See the section "How does it
-    work" in te docs for specifics.
+    It cleans any previously found tests, i.e., only tests defined in the
+    current cell are executed. To disable this behavior, use
+    [`ipytest.config(clean=False)`][ipytest.config].
 
-    For example:
+    Any arguments passed on the magic line are interpreted as command line
+    arguments to to pytest. For example calling the magic as
 
     ```python
     %%ipytest -qq
-
-
-    def test_example():
-        ...
-
     ```
+
+    is equivalent to passing `-qq` to pytest. See also the section ["How does it
+    work"](#how-does-it-work) for further details.
+
+    The keyword arguments passed to [`ipytest.run()`][ipytest.run] can be
+    customized by including a comment of the form `# ipytest: arg1=value1,
+    arg=value2` in the cell source. For example:
+
+    ```python
+    %%ipytest {MODULE}::test1
+    # ipytest: defopts=False
+    ```
+
+    is equivalent to `ipytest.run("{MODULE}::test1", defopts=False)`. In this
+    case, it deactivates default arguments and then instructs pytest to only
+    execute `test1`.
     """
-    if current_config["clean"] is not False:
-        clean_tests(current_config["clean"])
+    run_args = shlex.split(line)
+    run_kwargs = eval_run_kwargs(cell, module=module)
+
+    clean_tests(module=run_kwargs.get("module"))
 
     try:
         get_ipython().run_cell(cell)
@@ -103,10 +142,10 @@ def pytest_magic(line, cell):
         else:
             raise e
 
-    run(*shlex.split(line))
+    run(*run_args, **run_kwargs)
 
 
-def clean_tests(pattern=default_clean, items=None):
+def clean_tests(pattern=default, *, module=None):
     """Delete tests with names matching the given pattern.
 
     In IPython the results of all evaluations are kept in global variables
@@ -115,20 +154,26 @@ def clean_tests(pattern=default_clean, items=None):
     aims to simply this process.
 
     An effective pattern is to start with the cell containing tests with a call
-    to `clean_tests`, then defined all test cases, and finally call `run_tests`.
-    This way renaming tests works as expected.
+    to [`ipytest.clean_tests()`][ipytest.clean_tests], then defined all test
+    cases, and finally call [`ipytest.run()`][ipytest.run]. This way renaming
+    tests works as expected.
 
     **Parameters:**
 
-    - `pattern`: a glob pattern used to match the tests to delete.
+    - `pattern`: a glob pattern used to match the tests to delete. If not given,
+      the `"clean"` config option is used.
     - `items`: the globals object containing the tests. If `None` is given, the
         globals object is determined from the call stack.
     """
-    if items is None:
-        import __main__
+    pattern = default.unwrap(pattern, current_config["clean"])
 
-        items = vars(__main__)
+    if pattern is False:
+        return
 
+    if module is None:
+        import __main__ as module
+
+    items = vars(module)
     to_delete = [key for key in items.keys() if fnmatch.fnmatchcase(key, pattern)]
 
     for key in to_delete:
@@ -144,32 +189,32 @@ def reload(*mods):
     Usage:
 
     ```python
-    reload("ipytest._util", "ipytest")
+    ipytest.reload("ipytest._util", "ipytest")
     ```
     """
     for mod in mods:
         importlib.reload(importlib.import_module(mod))
 
 
-def _run_impl(*args, module, plugins):
-    with _prepared_env(module) as filename:
-        full_args = _build_full_args(args, filename)
+def _run_impl(*args, module, plugins, addopts, defopts, display_columns):
+    with _prepared_env(module, display_columns=display_columns) as filename:
+        full_args = _build_full_args(args, filename, addopts=addopts, defopts=defopts)
         return pytest.main(full_args, plugins=[*plugins, FixProgramNamePlugin()])
 
 
-def _build_full_args(args, filename):
+def _build_full_args(args, filename, *, addopts, defopts):
     def _fmt(arg):
         return arg.format(MODULE=filename)
 
     return [
-        *(_fmt(arg) for arg in current_config["addopts"]),
+        *(_fmt(arg) for arg in addopts),
         *(_fmt(arg) for arg in args),
-        *([filename] if current_config["defopts"] else []),
+        *([filename] if defopts else []),
     ]
 
 
 @contextlib.contextmanager
-def _prepared_env(module):
+def _prepared_env(module, *, display_columns):
     if module is None:  # pragma: no cover
         import __main__ as module
 
@@ -193,7 +238,7 @@ def _prepared_env(module):
 
         with patch(module, "__file__", str(path)):
             with register_module(module, module_name):
-                with patched_columns():
+                with patched_columns(display_columns=display_columns):
                     yield str(path)
 
 
@@ -274,9 +319,7 @@ def register_module(obj, name):
 
 
 @contextlib.contextmanager
-def patched_columns():
-    display_columns = current_config["display_columns"]
-
+def patched_columns(*, display_columns):
     if not display_columns:
         yield
         return
@@ -294,11 +337,11 @@ def patched_columns():
         del os.environ["COLUMNS"]
 
 
-def run_direct(func, *args, **kwargs):
+def run_func_direct(func, *args, **kwargs):
     return func(*args, **kwargs)
 
 
-def run_in_thread(func, *args, **kwargs):
+def run_func_in_thread(func, *args, **kwargs):
     res = None
 
     def _thread():
@@ -314,3 +357,40 @@ def run_in_thread(func, *args, **kwargs):
 
 def is_valid_module_name(name):
     return all(c not in name for c in ".- ")
+
+
+RUN_OPTIONS_MARKER = "# ipytest:"
+
+
+def eval_run_kwargs(cell: str, module=None) -> Dict[str, Any]:
+    """Parse the `ipytest:` comment inside a cell
+
+    If the cell does not start with `# ipytest:` and empty dict is returned.
+    Otherwise, the rest of the line is evaluated as keyword args. Any references
+    to variables are evaluated to the module object. If not given it defaults to
+    `__main__`.
+
+    If the module is given and not overwritten inside the comment, it is also
+    returned as keyword argument.
+    """
+    if module is not None:
+        eval_module = module
+
+    else:
+        import __main__ as eval_module
+
+    lines = cell.splitlines()
+    if not lines:
+        return {}
+
+    first_line = lines[0]
+    if not first_line.startswith(RUN_OPTIONS_MARKER):
+        return {}
+
+    run_options = first_line[len(RUN_OPTIONS_MARKER) :]
+    kwargs = eval(f"dict({run_options!s})", eval_module.__dict__, eval_module.__dict__)
+
+    if "module" not in kwargs and module is not None:
+        kwargs["module"] = module
+
+    return kwargs