+++ Changes.md
@@ -3,7 +3,15 @@
 Note: development is tracked on the [`develop` branch](https://github.com/chmp/ipytest/tree/develop).
 
 - `develop`:
-    - Use markdown for module documentation 
+    - Use markdown for module documentation
+    - Updated documentation (updated doc strings for run, ipytest magic)
+    - `ipytest.clean()` now expects a module as its argument
+    - Allow to specify all keyword arguments of `ipytest.run` also in
+      `%%ipytest` by including an initial comment of the form `# ipytest:
+      arg1=value1, arg2=value`
+    - Allow to override `addopts`, `defopts`, `run_in_thread`, `raise_on_error`,
+      `display_columns` in `ipytest.run` (this change also applies to
+      `%%ipytest` with the due to the previous change)
 - `0.12.0`:
     - Re-add the `raise_on_error` config option
     - Return the `exit_code` from `ipyest.run()`